moved helper functions to its own file
